refactor kuutamod.py for testcase
